+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+* fix(sand agent cli): add a timeout to the curl http client
+* fix(sand client): disable HTTP keepalives to prevent leaking file descriptors
 * chore(go): use go 1.17
 * chore(deps): bump github.com/Scalingo/go-handlers from 1.4.0 to 1.4.4
 * chore(deps): bump github.com/Scalingo/go-utils/etcd from 1.1.0 to 1.1.1